Implemented elevated privleges for prefs page.
Elevated privleges are now required to access the preferences page. If the user
tries to access the page with expired elevated privleges then they will be
asked to enter their password again. Their elevated privleges will be renewed
for 15 minutes upon entering the correct password and for each subsequent
request on the preferences page.